Move UI files to separate folder
